+++ src/debugger/calva-debug.ts
@@ -251,8 +251,6 @@ class CalvaDebugSession extends LoggingDebugSession {
             cljSession.sendDebugInput(':quit', id, key);
         }
 
-        debugDecorations.triggerUpdateDecorations();
-
         this.sendResponse(response);
     }
 
@@ -369,6 +367,13 @@ async function initializeDebugger(cljSession: NReplSession): Promise<void> {
     debugDecorations.activate();
 }
 
+function terminateDebugSession(): void {
+    if (vscode.debug.activeDebugSession) {
+        vscode.debug.activeDebugSession.customRequest(REQUESTS.SEND_TERMINATED_EVENT);
+    }
+    debugDecorations.triggerUpdateAndRenderDecorations();
+}
+
 export {
     CALVA_DEBUG_CONFIGURATION,
     DEBUG_ANALYTICS,
@@ -380,5 +385,6 @@ export {
     CalvaDebugAdapterDescriptorFactory,
     handleNeedDebugInput,
     initializeDebugger,
-    onNreplMessage
+    onNreplMessage,
+    terminateDebugSession
 };